DEV-3: Rename and refactor Boss class
- Rename Boss class
- Refactor Boss class to implement new Entity interface props
- Make Boss class abstract
