--- Graduation_Game/Entities/Bosses/BossOne.cs
@@ -1,25 +1,0 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Graduation_Game.Entities.Bosses
-{
-    class BossOne : Entity
-    {
-        public Texture2D Sprite { get; set; }
-        public Vector2 Position { get; set; }
-        public int DrawOder { get; set; }
-
-        public void Draw(SpriteBatch spriteBatch, GameTime gameTime)
-        {
-            
-        }
-
-        public void Update(GameTime gameTime)
-        {
-            
-        }
-    }
-}
